Flytt Json-ting til egne filer
